Migrate the docs components to app/components to align with Phlex conversions
